requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docopt==0.6.2
Flask==1.1.2
future==0.18.2
graphviz==0.14
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
pipreqs==0.4.10
PyYAML==5.3.1
requests==2.24.0
torch==1.5.1
torchvision==0.6.1
urllib3==1.25.9
Werkzeug==1.0.1
yarg==0.1.9
